refactor(show): split the season-ordering controls into their own module
`season_components.ex` had reached 520 lines against the project's ~500 LOC
component guideline. `season_order_controls/1` and its three helpers move to a
sibling `SeasonOrderComponents`, leaving 434 and 102 lines. The two share a
page but not a subject: one is about which of TVDB's parallel orderings the
show is in, the other about rendering whichever one that turns out to be.

Refs #490
